Re-organize FE files by the page used on
