+++ react_front_end/src/pages/CreateChallengePage/CreateChallengeInputs.js
