+++ react_front_end/src/pages/CreateChallengePage/CreateChallengePrompt.js
