+++ react_front_end/src/pages/GameInfoPage/GameInfoButtonOptions.js
