+++ react_front_end/src/pages/GameInfoPage/NextChallenge.js
