+++ react_front_end/src/pages/GameInfoPage/NextTribalCouncil.js
