+++ react_front_end/src/pages/GameInfoPage/NumberOfPlayers.js
