+++ react_front_end/src/pages/JoinGamePage/JoinGameInputs.js
