+++ react_front_end/src/pages/JoinGamePage/JoinGamePrompt.js
