+++ react_front_end/src/pages/LandingPage/LandingPage.scss
