+++ react_front_end/src/pages/LandingPage/LandingPageInformation.js
