+++ react_front_end/src/pages/StartGamePage/StartGameInputs.js
