+++ react_front_end/src/pages/StartGamePage/StartGamePage.js
@@ -1,12 +1,12 @@
 import { makeStyles } from "@material-ui/core/styles";
 import React from "react";
-import AppTitle from "./components/AppTitle";
-import Footer from "./components/Footer";
-import StartGamePrompt from "./components/StartGamePrompt";
-import StartGameInputs from "./components/StartGameInputs";
-import TriangleLogo from "./components/TriangleLogo";
-import LogoContainer from "./components/LogoContainer";
-import { StatisticsLeft, StatisticsRight } from "./components/GameStatistics";
+import AppTitle from "../common/AppTitle";
+import Footer from "../common/Footer";
+import StartGamePrompt from "./StartGamePrompt";
+import StartGameInputs from "./StartGameInputs";
+import TriangleLogo from "../common/TriangleLogo";
+import LogoContainer from "../common/LogoContainer";
+import { StatisticsLeft, StatisticsRight } from "../common/GameStatistics";
 
 export default function StartGamePage() {
   const useStyles = makeStyles((theme) => ({
