+++ react_front_end/src/pages/StartGamePage/StartGamePrompt.js
