+++ react_front_end/src/pages/SubmitPage/ChallengeInfo.js
@@ -2,7 +2,7 @@ import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import Paper from "@material-ui/core/Paper";
 import Typography from "@material-ui/core/Typography";
-import { GameName } from "./GameName";
+import { GameName } from "../common/GameName";
 
 const useStyles = makeStyles((theme) => ({
   root: {
