+++ react_front_end/src/pages/common/AppTitle.js
