+++ react_front_end/src/pages/common/Footer.scss
