+++ react_front_end/src/pages/common/GameName.js
