+++ react_front_end/src/pages/common/GameStatistics.scss
