+++ react_front_end/src/pages/common/LogoContainer.scss
