+++ react_front_end/src/pages/common/Preloader.scss
