+++ react_front_end/src/pages/common/TriangleLogo.js
