--- react_front_end/src/pages/components/AppBar.js
@@ -1,40 +1,0 @@
-import React from "react";
-import { makeStyles } from "@material-ui/core/styles";
-import AppBar from "@material-ui/core/AppBar";
-import Toolbar from "@material-ui/core/Toolbar";
-import IconButton from "@material-ui/core/IconButton";
-import MenuIcon from "@material-ui/icons/Menu";
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 0,
-  },
-  menuButton: {
-    marginLeft: theme.spacing(2),
-  },
-  toolbar: {
-    justifyContent: "flex-end",
-    height: "10vh",
-  },
-}));
-
-export default function ButtonAppBar() {
-  const classes = useStyles();
-
-  return (
-    <div className={classes.root}>
-      <AppBar position="sticky">
-        <Toolbar className={classes.toolbar}>
-          <IconButton
-            edge="start"
-            className={classes.menuButton}
-            color="inherit"
-            aria-label="menu"
-          >
-            <MenuIcon />
-          </IconButton>
-        </Toolbar>
-      </AppBar>
-    </div>
-  );
-}
