--- react_front_end/src/pages/components/Button.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-export const Button = React.forwardRef(({ label, click }, ref) => (
-  <button ref={ref} onClick={click}>
-    {label}
-  </button>
-));
