--- react_front_end/src/pages/components/LandingPageHeaderLogoSvg.js
@@ -1,27 +1,0 @@
-import { withStyles } from "@material-ui/core";
-import { makeStyles } from "@material-ui/core/styles";
-import React from "react";
-import { ReactComponent as Svg } from "./svg/vir-us.logo.svg";
-import "./LandingPageHeaderLogoSvg.scss";
-import LandingPageSocialMediaButtons from "./LandingPageSocialMediaButtons";
-
-const useStyles = makeStyles((theme) => ({
-  root: {},
-}));
-
-export const Logo = () => {
-  return <Svg />;
-};
-
-const LandingPageHeaderLogoSvg = () => {
-  return (
-    <header className="header">
-      <div className="logo-container">
-        <Logo />
-      </div>
-      <LandingPageSocialMediaButtons />
-    </header>
-  );
-};
-
-export default withStyles(useStyles)(LandingPageHeaderLogoSvg);
