--- react_front_end/src/pages/components/LandingPageSocialMediaIcons.scss
@@ -1,36 +1,0 @@
-.social-media {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-evenly;
-  min-width: 260px;
-
-  .social-media-button {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-
-    margin-top: 2em;
-
-    border-radius: 5px;
-    background: linear-gradient(-70deg, #ffffff, #d7d7d7);
-    box-shadow: -3px -3px 6px #cbcbcb, 10px 10px 25px #ffffff;
-
-    svg {
-      font-size: 3em;
-      fill: rgb(0, 0, 0);
-
-      // default ui state
-      filter: blur(0.2rem);
-
-      // ui transition
-      animation-name: ui-component-blur-up;
-      animation-fill-mode: forwards;
-      animation-duration: 1s;
-      animation-timing-function: cubic-bezier(0.55, 0.085, 0.68, 0.53);
-      animation-delay: 0.5s;
-      // default ui state
-      opacity: 0;
-    }
-  }
-}
